Extract IPP integration as HAL function
