+++ modules/imgproc/src/color.hpp
@@ -333,184 +333,6 @@ struct OclHelper
 
 #endif
 
-
-
-#if defined (HAVE_IPP) && (IPP_VERSION_X100 >= 700)
-#  define NEED_IPP 1
-#else
-#  define NEED_IPP 0
-#endif
-
-#if NEED_IPP
-
-#define MAX_IPP8u   255
-#define MAX_IPP16u  65535
-#define MAX_IPP32f  1.0
-
-typedef IppStatus (CV_STDCALL* ippiReorderFunc)(const void *, int, void *, int, IppiSize, const int *);
-typedef IppStatus (CV_STDCALL* ippiGeneralFunc)(const void *, int, void *, int, IppiSize);
-typedef IppStatus (CV_STDCALL* ippiColor2GrayFunc)(const void *, int, void *, int, IppiSize, const Ipp32f *);
-
-template <typename Cvt>
-class CvtColorIPPLoop_Invoker :
-        public ParallelLoopBody
-{
-public:
-
-    CvtColorIPPLoop_Invoker(const uchar * src_data_, size_t src_step_, uchar * dst_data_, size_t dst_step_, int width_, const Cvt& _cvt, bool *_ok) :
-        ParallelLoopBody(), src_data(src_data_), src_step(src_step_), dst_data(dst_data_), dst_step(dst_step_), width(width_), cvt(_cvt), ok(_ok)
-    {
-        *ok = true;
-    }
-
-    virtual void operator()(const Range& range) const CV_OVERRIDE
-    {
-        const void *yS = src_data + src_step * range.start;
-        void *yD = dst_data + dst_step * range.start;
-        if( !cvt(yS, static_cast<int>(src_step), yD, static_cast<int>(dst_step), width, range.end - range.start) )
-            *ok = false;
-        else
-        {
-            CV_IMPL_ADD(CV_IMPL_IPP|CV_IMPL_MT);
-        }
-    }
-
-private:
-    const uchar * src_data;
-    const size_t src_step;
-    uchar * dst_data;
-    const size_t dst_step;
-    const int width;
-    const Cvt& cvt;
-    bool *ok;
-
-    const CvtColorIPPLoop_Invoker& operator= (const CvtColorIPPLoop_Invoker&);
-};
-
-
-template <typename Cvt>
-bool CvtColorIPPLoop(const uchar * src_data, size_t src_step, uchar * dst_data, size_t dst_step, int width, int height, const Cvt& cvt)
-{
-    bool ok;
-    parallel_for_(Range(0, height), CvtColorIPPLoop_Invoker<Cvt>(src_data, src_step, dst_data, dst_step, width, cvt, &ok), (width * height)/(double)(1<<16) );
-    return ok;
-}
-
-
-template <typename Cvt>
-bool CvtColorIPPLoopCopy(const uchar * src_data, size_t src_step, int src_type, uchar * dst_data, size_t dst_step, int width, int height, const Cvt& cvt)
-{
-    Mat temp;
-    Mat src(Size(width, height), src_type, const_cast<uchar*>(src_data), src_step);
-    Mat source = src;
-    if( src_data == dst_data )
-    {
-        src.copyTo(temp);
-        source = temp;
-    }
-    bool ok;
-    parallel_for_(Range(0, source.rows),
-                  CvtColorIPPLoop_Invoker<Cvt>(source.data, source.step, dst_data, dst_step,
-                                               source.cols, cvt, &ok),
-                  source.total()/(double)(1<<16) );
-    return ok;
-}
-
-
-struct IPPGeneralFunctor
-{
-    IPPGeneralFunctor(ippiGeneralFunc _func) : ippiColorConvertGeneral(_func){}
-    bool operator()(const void *src, int srcStep, void *dst, int dstStep, int cols, int rows) const
-    {
-        return ippiColorConvertGeneral ? CV_INSTRUMENT_FUN_IPP(ippiColorConvertGeneral, src, srcStep, dst, dstStep, ippiSize(cols, rows)) >= 0 : false;
-    }
-private:
-    ippiGeneralFunc ippiColorConvertGeneral;
-};
-
-
-struct IPPReorderFunctor
-{
-    IPPReorderFunctor(ippiReorderFunc _func, int _order0, int _order1, int _order2) : ippiColorConvertReorder(_func)
-    {
-        order[0] = _order0;
-        order[1] = _order1;
-        order[2] = _order2;
-        order[3] = 3;
-    }
-    bool operator()(const void *src, int srcStep, void *dst, int dstStep, int cols, int rows) const
-    {
-        return ippiColorConvertReorder ? CV_INSTRUMENT_FUN_IPP(ippiColorConvertReorder, src, srcStep, dst, dstStep, ippiSize(cols, rows), order) >= 0 : false;
-    }
-private:
-    ippiReorderFunc ippiColorConvertReorder;
-    int order[4];
-};
-
-
-struct IPPReorderGeneralFunctor
-{
-    IPPReorderGeneralFunctor(ippiReorderFunc _func1, ippiGeneralFunc _func2, int _order0, int _order1, int _order2, int _depth) :
-        ippiColorConvertReorder(_func1), ippiColorConvertGeneral(_func2), depth(_depth)
-    {
-        order[0] = _order0;
-        order[1] = _order1;
-        order[2] = _order2;
-        order[3] = 3;
-    }
-    bool operator()(const void *src, int srcStep, void *dst, int dstStep, int cols, int rows) const
-    {
-        if (ippiColorConvertReorder == 0 || ippiColorConvertGeneral == 0)
-            return false;
-
-        Mat temp;
-        temp.create(rows, cols, CV_MAKETYPE(depth, 3));
-        if(CV_INSTRUMENT_FUN_IPP(ippiColorConvertReorder, src, srcStep, temp.ptr(), (int)temp.step[0], ippiSize(cols, rows), order) < 0)
-            return false;
-        return CV_INSTRUMENT_FUN_IPP(ippiColorConvertGeneral, temp.ptr(), (int)temp.step[0], dst, dstStep, ippiSize(cols, rows)) >= 0;
-    }
-private:
-    ippiReorderFunc ippiColorConvertReorder;
-    ippiGeneralFunc ippiColorConvertGeneral;
-    int order[4];
-    int depth;
-};
-
-
-struct IPPGeneralReorderFunctor
-{
-    IPPGeneralReorderFunctor(ippiGeneralFunc _func1, ippiReorderFunc _func2, int _order0, int _order1, int _order2, int _depth) :
-        ippiColorConvertGeneral(_func1), ippiColorConvertReorder(_func2), depth(_depth)
-    {
-        order[0] = _order0;
-        order[1] = _order1;
-        order[2] = _order2;
-        order[3] = 3;
-    }
-    bool operator()(const void *src, int srcStep, void *dst, int dstStep, int cols, int rows) const
-    {
-        if (ippiColorConvertGeneral == 0 || ippiColorConvertReorder == 0)
-            return false;
-
-        Mat temp;
-        temp.create(rows, cols, CV_MAKETYPE(depth, 3));
-        if(CV_INSTRUMENT_FUN_IPP(ippiColorConvertGeneral, src, srcStep, temp.ptr(), (int)temp.step[0], ippiSize(cols, rows)) < 0)
-            return false;
-        return CV_INSTRUMENT_FUN_IPP(ippiColorConvertReorder, temp.ptr(), (int)temp.step[0], dst, dstStep, ippiSize(cols, rows), order) >= 0;
-    }
-private:
-    ippiGeneralFunc ippiColorConvertGeneral;
-    ippiReorderFunc ippiColorConvertReorder;
-    int order[4];
-    int depth;
-};
-
-extern ippiReorderFunc ippiSwapChannelsC3C4RTab[CV_DEPTH_MAX];
-extern ippiReorderFunc ippiSwapChannelsC4C3RTab[CV_DEPTH_MAX];
-extern ippiReorderFunc ippiSwapChannelsC3RTab[CV_DEPTH_MAX];
-
-#endif
-
 #ifdef HAVE_OPENCL
 
 bool oclCvtColorBGR2Luv( InputArray _src, OutputArray _dst, int bidx, bool srgb );
