+++ modules/imgproc/src/color_lab.cpp
@@ -4051,47 +4051,6 @@ struct Luv2RGB_b
 };
 
 //
-// IPP functions
-//
-
-#if NEED_IPP
-
-#if !IPP_DISABLE_RGB_XYZ
-static ippiGeneralFunc ippiRGB2XYZTab[CV_DEPTH_MAX] =
-{
-    (ippiGeneralFunc)ippiRGBToXYZ_8u_C3R, 0, (ippiGeneralFunc)ippiRGBToXYZ_16u_C3R, 0,
-    0, (ippiGeneralFunc)ippiRGBToXYZ_32f_C3R, 0, 0
-};
-#endif
-
-#if !IPP_DISABLE_XYZ_RGB
-static ippiGeneralFunc ippiXYZ2RGBTab[CV_DEPTH_MAX] =
-{
-    (ippiGeneralFunc)ippiXYZToRGB_8u_C3R, 0, (ippiGeneralFunc)ippiXYZToRGB_16u_C3R, 0,
-    0, (ippiGeneralFunc)ippiXYZToRGB_32f_C3R, 0, 0
-};
-#endif
-
-#if !IPP_DISABLE_RGB_LAB
-static ippiGeneralFunc ippiRGBToLUVTab[CV_DEPTH_MAX] =
-{
-    (ippiGeneralFunc)ippiRGBToLUV_8u_C3R, 0, (ippiGeneralFunc)ippiRGBToLUV_16u_C3R, 0,
-    0, (ippiGeneralFunc)ippiRGBToLUV_32f_C3R, 0, 0
-};
-#endif
-
-#if !IPP_DISABLE_LAB_RGB
-static ippiGeneralFunc ippiLUVToRGBTab[CV_DEPTH_MAX] =
-{
-    (ippiGeneralFunc)ippiLUVToRGB_8u_C3R, 0, (ippiGeneralFunc)ippiLUVToRGB_16u_C3R, 0,
-    0, (ippiGeneralFunc)ippiLUVToRGB_32f_C3R, 0, 0
-};
-#endif
-
-#endif
-
-
-//
 // HAL functions
 //
 
@@ -4106,38 +4065,6 @@ void cvtBGRtoXYZ(const uchar * src_data, size_t src_step,
     CV_INSTRUMENT_REGION();
 
     CALL_HAL(cvtBGRtoXYZ, cv_hal_cvtBGRtoXYZ, src_data, src_step, dst_data, dst_step, width, height, depth, scn, swapBlue);
-
-#if defined(HAVE_IPP) && IPP_VERSION_X100 >= 700
-#if !IPP_DISABLE_RGB_XYZ
-    CV_IPP_CHECK()
-    {
-        if(scn == 3 && depth != CV_32F && !swapBlue)
-        {
-            if( CvtColorIPPLoopCopy(src_data, src_step, CV_MAKETYPE(depth, scn), dst_data, dst_step, width, height,
-                                    IPPReorderGeneralFunctor(ippiSwapChannelsC3RTab[depth], ippiRGB2XYZTab[depth], 2, 1, 0, depth)) )
-                return;
-        }
-        else if(scn == 4 && depth != CV_32F && !swapBlue)
-        {
-            if( CvtColorIPPLoop(src_data, src_step, dst_data, dst_step, width, height,
-                                IPPReorderGeneralFunctor(ippiSwapChannelsC4C3RTab[depth], ippiRGB2XYZTab[depth], 2, 1, 0, depth)) )
-                return;
-        }
-        else if(scn == 3 && depth != CV_32F && swapBlue)
-        {
-            if( CvtColorIPPLoopCopy(src_data, src_step, CV_MAKETYPE(depth, scn), dst_data, dst_step, width, height,
-                                    IPPGeneralFunctor(ippiRGB2XYZTab[depth])) )
-                return;
-        }
-        else if(scn == 4 && depth != CV_32F && swapBlue)
-        {
-            if( CvtColorIPPLoop(src_data, src_step, dst_data, dst_step, width, height,
-                                IPPReorderGeneralFunctor(ippiSwapChannelsC4C3RTab[depth], ippiRGB2XYZTab[depth], 0, 1, 2, depth)) )
-                return;
-        }
-    }
-#endif
-#endif
 
     int blueIdx = swapBlue ? 2 : 0;
     if( depth == CV_8U )
@@ -4158,38 +4085,6 @@ void cvtXYZtoBGR(const uchar * src_data, size_t src_step,
 
     CALL_HAL(cvtXYZtoBGR, cv_hal_cvtXYZtoBGR, src_data, src_step, dst_data, dst_step, width, height, depth, dcn, swapBlue);
 
-#if defined(HAVE_IPP) && IPP_VERSION_X100 >= 700
-#if !IPP_DISABLE_XYZ_RGB
-    CV_IPP_CHECK()
-    {
-        if(dcn == 3 && depth != CV_32F && !swapBlue)
-        {
-            if( CvtColorIPPLoopCopy(src_data, src_step, CV_MAKETYPE(depth, 3), dst_data, dst_step, width, height,
-                                    IPPGeneralReorderFunctor(ippiXYZ2RGBTab[depth], ippiSwapChannelsC3RTab[depth], 2, 1, 0, depth)) )
-                return;
-        }
-        else if(dcn == 4 && depth != CV_32F && !swapBlue)
-        {
-            if( CvtColorIPPLoop(src_data, src_step, dst_data, dst_step, width, height,
-                                IPPGeneralReorderFunctor(ippiXYZ2RGBTab[depth], ippiSwapChannelsC3C4RTab[depth], 2, 1, 0, depth)) )
-                return;
-        }
-        if(dcn == 3 && depth != CV_32F && swapBlue)
-        {
-            if( CvtColorIPPLoopCopy(src_data, src_step, CV_MAKETYPE(depth, 3), dst_data, dst_step, width, height,
-                                    IPPGeneralFunctor(ippiXYZ2RGBTab[depth])) )
-                return;
-        }
-        else if(dcn == 4 && depth != CV_32F && swapBlue)
-        {
-            if( CvtColorIPPLoop(src_data, src_step, dst_data, dst_step, width, height,
-                                IPPGeneralReorderFunctor(ippiXYZ2RGBTab[depth], ippiSwapChannelsC3C4RTab[depth], 0, 1, 2, depth)) )
-                return;
-        }
-    }
-#endif
-#endif
-
     int blueIdx = swapBlue ? 2 : 0;
     if( depth == CV_8U )
         CvtColorLoop(src_data, src_step, dst_data, dst_step, width, height, XYZ2RGB_i<uchar>(dcn, blueIdx, 0));
@@ -4209,75 +4104,6 @@ void cvtBGRtoLab(const uchar * src_data, size_t src_step,
     CV_INSTRUMENT_REGION();
 
     CALL_HAL(cvtBGRtoLab, cv_hal_cvtBGRtoLab, src_data, src_step, dst_data, dst_step, width, height, depth, scn, swapBlue, isLab, srgb);
-
-#if defined(HAVE_IPP) && !IPP_DISABLE_RGB_LAB
-    CV_IPP_CHECK()
-    {
-        if (!srgb)
-        {
-            if (isLab)
-            {
-                if (scn == 3 && depth == CV_8U && !swapBlue)
-                {
-                    if (CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralFunctor((ippiGeneralFunc)ippiBGRToLab_8u_C3R)))
-                        return;
-                }
-                else if (scn == 4 && depth == CV_8U && !swapBlue)
-                {
-                    if (CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPReorderGeneralFunctor(ippiSwapChannelsC4C3RTab[depth],
-                                                                 (ippiGeneralFunc)ippiBGRToLab_8u_C3R, 0, 1, 2, depth)))
-                        return;
-                }
-                else if (scn == 3 && depth == CV_8U && swapBlue) // slower than OpenCV
-                {
-                    if (CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPReorderGeneralFunctor(ippiSwapChannelsC3RTab[depth],
-                                                                 (ippiGeneralFunc)ippiBGRToLab_8u_C3R, 2, 1, 0, depth)))
-                        return;
-                }
-                else if (scn == 4 && depth == CV_8U && swapBlue) // slower than OpenCV
-                {
-                    if (CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPReorderGeneralFunctor(ippiSwapChannelsC4C3RTab[depth],
-                                                                 (ippiGeneralFunc)ippiBGRToLab_8u_C3R, 2, 1, 0, depth)))
-                        return;
-                }
-            }
-            else
-            {
-                if (scn == 3 && swapBlue)
-                {
-                    if (CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralFunctor(ippiRGBToLUVTab[depth])))
-                        return;
-                }
-                else if (scn == 4 && swapBlue)
-                {
-                    if (CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPReorderGeneralFunctor(ippiSwapChannelsC4C3RTab[depth],
-                                                                 ippiRGBToLUVTab[depth], 0, 1, 2, depth)))
-                        return;
-                }
-                else if (scn == 3 && !swapBlue)
-                {
-                    if (CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPReorderGeneralFunctor(ippiSwapChannelsC3RTab[depth],
-                                                                 ippiRGBToLUVTab[depth], 2, 1, 0, depth)))
-                        return;
-                }
-                else if (scn == 4 && !swapBlue)
-                {
-                    if (CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPReorderGeneralFunctor(ippiSwapChannelsC4C3RTab[depth],
-                                                                 ippiRGBToLUVTab[depth], 2, 1, 0, depth)))
-                        return;
-                }
-            }
-        }
-    }
-#endif
 
     int blueIdx = swapBlue ? 2 : 0;
     if(isLab)
@@ -4306,75 +4132,6 @@ void cvtLabtoBGR(const uchar * src_data, size_t src_step,
     CV_INSTRUMENT_REGION();
 
     CALL_HAL(cvtLabtoBGR, cv_hal_cvtLabtoBGR, src_data, src_step, dst_data, dst_step, width, height, depth, dcn, swapBlue, isLab, srgb);
-
-#if defined(HAVE_IPP) && !IPP_DISABLE_LAB_RGB
-    CV_IPP_CHECK()
-    {
-        if (!srgb)
-        {
-            if (isLab)
-            {
-                if( dcn == 3 && depth == CV_8U && !swapBlue)
-                {
-                    if( CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralFunctor((ippiGeneralFunc)ippiLabToBGR_8u_C3R)) )
-                        return;
-                }
-                else if( dcn == 4 && depth == CV_8U && !swapBlue)
-                {
-                    if( CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralReorderFunctor((ippiGeneralFunc)ippiLabToBGR_8u_C3R,
-                                                                 ippiSwapChannelsC3C4RTab[depth], 0, 1, 2, depth)) )
-                        return;
-                }
-                if( dcn == 3 && depth == CV_8U && swapBlue)
-                {
-                    if( CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralReorderFunctor((ippiGeneralFunc)ippiLabToBGR_8u_C3R,
-                                                                 ippiSwapChannelsC3RTab[depth], 2, 1, 0, depth)) )
-                        return;
-                }
-                else if( dcn == 4 && depth == CV_8U && swapBlue)
-                {
-                    if( CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralReorderFunctor((ippiGeneralFunc)ippiLabToBGR_8u_C3R,
-                                                                 ippiSwapChannelsC3C4RTab[depth], 2, 1, 0, depth)) )
-                        return;
-                }
-            }
-            else
-            {
-                if( dcn == 3 && swapBlue)
-                {
-                    if( CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralFunctor(ippiLUVToRGBTab[depth])) )
-                        return;
-                }
-                else if( dcn == 4 && swapBlue)
-                {
-                    if( CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralReorderFunctor(ippiLUVToRGBTab[depth],
-                                                                 ippiSwapChannelsC3C4RTab[depth], 0, 1, 2, depth)) )
-                        return;
-                }
-                if( dcn == 3 && !swapBlue)
-                {
-                    if( CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralReorderFunctor(ippiLUVToRGBTab[depth],
-                                                                 ippiSwapChannelsC3RTab[depth], 2, 1, 0, depth)) )
-                        return;
-                }
-                else if( dcn == 4 && !swapBlue)
-                {
-                    if( CvtColorIPPLoop(src_data, src_step, dst_data,dst_step, width, height,
-                                        IPPGeneralReorderFunctor(ippiLUVToRGBTab[depth],
-                                                                 ippiSwapChannelsC3C4RTab[depth], 2, 1, 0, depth)) )
-                        return;
-                }
-            }
-        }
-    }
-#endif
 
     int blueIdx = swapBlue ? 2 : 0;
     if(isLab)
